Comitting a bunch of changes.
Todo:
- parallax scrolling
- finish replay function for checkpoint based stages
- bug fix for Orbital Station
- name change for Orbital Station

+++ src/game/essentials/SoundBank.java
@@ -267,22 +267,22 @@ public class SoundBank
 	/**
 	 * Returns the volume calculated by the distance between the two given points. 
 	 */
-	public static float getVolume(float emitterX, float emitterY, float listenerX, float listenerY, float maxDistance, float maxVolume, float power)
+	public static double getVolume(float emitterX, float emitterY, float listenerX, float listenerY, float maxDistance, float maxVolume, float power)
 	{
 		double distance = Fundementals.distance(emitterX, emitterY, listenerX, listenerY);
 		double candidate = power * Math.max((1 / Math.sqrt(distance)) - (1 / Math.sqrt(maxDistance)), 0);
 		
-		return (float) Math.min(candidate, maxVolume);
+		return Math.min(candidate, maxVolume);
 	}
 	
 	/**
 	 * Returns the volume calculated by the distance between the two {@code GameObjects}. 
 	 */
-	public static float getVolume(GameObject emitter, GameObject listener, float maxDistance, float maxVolume, float power)
+	public static double getVolume(GameObject emitter, GameObject listener, float maxDistance, float maxVolume, float power)
 	{
 		double distance = Fundementals.distance(emitter, listener);
 		double candidate = power * Math.max((1 / Math.sqrt(distance)) - (1 / Math.sqrt(maxDistance)), 0);
 		
-		return (float) Math.min(candidate, maxVolume);
+		return Math.min(candidate, maxVolume);
 	}
 }
